feat(snes9x): MK3 accuracy fixes, RAM placement, and always-on frame telemetry
Accuracy (the MK3 announcer fix, verified byte-identical to the 1.6x
reference for 136.9 s and confirmed on hardware):

  - getset.c: perform the bus access BEFORE charging its cycles, and charge
    nothing at all while CPU.InDMA. Route 16-bit MAP_PPU accesses through the
    byte path so $2140-3 sees two separate bus cycles.
  - cpu.c: reset enters the event ring at RENDER with Cycles=182, SL=-3, and
    the second init block no longer discards the vector fetch's 24 cycles.
  - cpuexec.c: H/V timer IRQ as an absolute Timings.NextIRQTimer deadline
    tested per opcode, rebased by H_Max each scanline; retire stale
    IRQ_PENDING before waking WAI; re-check NMI after a WAI wake.
  - cpuaddr.h: charge the direct-page penalty on all eight dp modes.
  - cpuops.c: interrupt entry costs MemSpeed + ONE_CYCLE up front; OpDC no
    longer charges two cycles the 65816 does not spend.
  - ppu.c/.h: HBlankStart is 1096 (H=274), $4212 gains the < HBlankEnd term,
    $2137 returns OpenBus1, and $2139/a use 1.6x's VRAM prefetch latch.
  - dma.c: HDMA restructured to 1.6x's shape (HDMAReadLineCount, two-pass
    S9xDoHDMA); $43x5/$43x6 is one register pair, as a union.
  - memmap.c: SRAM initialises to 0x60, not zero.

Performance, all verified byte-exact offline (MK3 announcer gate 1.000,
full-frame video hash unchanged over 2,500 frames, SMW/Axelay/Batman Forever
render identical):

  - dma.c: iterate armed HDMA channels via __builtin_ctz instead of testing
    all eight twice per scanline (3,172 -> 1,128 iterations/frame).
  - getset.c: split the bus accessors into hot/cold halves so the 91.6%
    direct-pointer path does not carry the switch's register pressure; the
    cold halves are not BUS_HOT, so they land in flash. Read MapInfo[block]
    once on the fast path.
  - cpuexec.c: accumulate the instruction count in a register and flush it
    once per frame instead of a volatile RMW per emulated instruction.
  - gfx.c: hoist DrawTilePtr, BG.TileSize, IPPU.HalfWidthPixels and
    GFX.PixSize out of DrawBackground's middle-tile loops.

  Measured on hardware, these are worth nothing (773,720 vs 772,662
  instr/s, inside noise). They are kept because they are byte-exact and
  free 944 bytes of RAM, not because they are faster. The real bottleneck
  is the renderer - see below.

Telemetry: the FRANK_SNES_PROFILE build does not boot on the C2, so the
ship build now carries its own always-on timers (ship_emul_us, ship_upd_us,
ship_rs_us, ship_rs_sub_us, tile call counts). These are what established
the actual frame budget: emulation 17.4 ms/frame, of which S9xUpdateScreen
is 46% and main-screen tile drawing alone is 28%. The interpreter is the
smaller half. That measurement is the basis for moving the PPU to the C2
slave, which has 373 KB of RAM and an entirely unused PSRAM.

+++ src/snes9x/spc700_blargg.c
@@ -1845,12 +1845,21 @@ static uint32_t spc_remainder;
 static uint32_t ratio_numerator   = APU_NUMERATOR_NTSC;
 static uint32_t ratio_denominator = APU_DENOMINATOR_NTSC;
 
+/* 64-bit intermediate, deliberately.
+ *
+ * CPU.Cycles is normally under one scanline, but a long DMA runs with the
+ * event drain suppressed and can leave it a couple of hundred thousand cycles
+ * ahead before the drain unwinds it. ratio_numerator times that span exceeds
+ * 32 bits - measured offline at 34176 * 263216 = 8,995,670,016, which wraps to
+ * 405,735,424 and yields 572 APU clocks where 12,678 were owed. The SPC700
+ * silently lost ~12,000 clocks every time it happened and ended up
+ * permanently out of phase with the CPU. */
 #define S9X_APU_GET_CLOCK(cpucycles) \
-   ((int32_t)((ratio_numerator * (uint32_t)((cpucycles) - reference_time) \
+   ((int32_t)(((uint64_t)ratio_numerator * (uint32_t)((cpucycles) - reference_time) \
                + spc_remainder) / ratio_denominator))
 #define S9X_APU_GET_CLOCK_REMAINDER(cpucycles) \
-   ((ratio_numerator * (uint32_t)((cpucycles) - reference_time) \
-     + spc_remainder) % ratio_denominator)
+   ((uint32_t)(((uint64_t)ratio_numerator * (uint32_t)((cpucycles) - reference_time) \
+     + spc_remainder) % ratio_denominator))
 
 bool S9xInitAPU(void)
 {
